day-26-28/flask-unit-testing/requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
coverage==7.2.7
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
